requirements.txt
pyautogui==0.9.54
opencv-python==4.8.0.76
numpy<2
google-cloud-vision==3.4.4
google-cloud-translate==3.11.1
pillow>=10.0.0
customtkinter>=5.2.0
scikit-image>=0.19.0
PyQt5>=5.15.0
requests>=2.31.0
pytesseract>=0.3.10
paddlepaddle>=2.5.0
paddleocr>=2.7.0  # Supports PP-OCRv4 model for improved accuracy
importlib-metadata>=4.0.0; python_version<"3.10"